example/schemas/requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
fastavro==1.7.1
idna==3.4
kafka-python==2.0.2
requests==2.28.2
urllib3==1.26.14
